pylbo: refactor file_handler > file_loader
