+++ post_processing/pylbo/__init__.py
@@ -1,7 +1,7 @@
 from pylbo._version import __version__
 from pylbo.automation.api import generate_parfiles, run_legolas
-from pylbo.file_handler import load, load_logfile, load_series
 from pylbo.utilities import logger
+from pylbo.utilities.datfiles.file_loader import load, load_logfile, load_series
 from pylbo.utilities.logger import disable_logging, set_loglevel
 from pylbo.visualisation.api import (
     plot_continua,
